391: move player role functions into look/playerRoles.js
